pnf_cli::clicommands
Function build_command_registry

pub fn build_command_registry() -> HashMap<&'static str, Command>

Builds and returns a HashMap of available commands, each represented by a Command structure.

This function initializes a registry of commands that can be executed in different modes (e.g., UserMode, PrivilegedMode, ConfigMode, etc.) within a router-like system. Each command is associated with a name, description, suggestions for usage, and an execution function that defines its behavior.

The commands registered include:

enable: Switches from User EXEC mode to Privileged EXEC mode.

enable secret: Sets a secret password for privileged EXEC mode access, using a stronger hash for security than the enable password command.

enable password: Configures a password for privileged EXEC mode access. This password is weaker than the enable secret and should be avoided when possible.

configure terminal: Enters Global Configuration mode.

interface: Enters Interface Configuration mode for a specified interface. Should enter the interface name as an input

interface range: Enters the Interface Configuration mode but for the entire range.

hostname: Changes the hostname of the device.

ifconfig: Displays or configures network details of the router.

exit: This command navigates through the modes in reverse order (eg: ConfigMode –> UserMode)

show: Displays all the show commands when specific command is passed

show running-config: Displays the current running configuration from a JSON file.

show startup-config: Displays the initial configuration settings stored in the NVRAM of a router, which are loaded upon booting the device.

show version: Displays the software version information.

show clock: Displays the current clock date and time.

show interfaces: Displays statistics for all interfaces, including a brief overview or detailed information.

show ip interfaces brief: Displays a summary of the router interfaces

show ip route: Displays the ip routes defined

show vlan: Displays information and status of VLANs.

show ip ospf neighbor: Displays information about OSPF neighbors, including their state, router ID, and the interface used for adjacency.

show access-lists: Displays the current configuration of all ACLs on the device, showing the list of ACL entries and their statistics (matches, actions, etc.).

show ntp associations: Displays the status of NTP associations with servers or clients, showing the synchronization status and other details.

show ntp: Displays information about the current NTP configuration, associations, and synchronization status.

show processes: Shows the system processes and memories

show uptime: Shows the time from the last reboot

show login: Shows the details about login delay

reload: Reloads the system

debug all: Debug all the processors

undebug all: Undebug all the processors

write memory: Saves the running configuration to the startup configuration.

copy running-config: Copies the running configuration to the startup configuration or to a new file if mentioned.

help: Displays a list of available commands.

clock set: Changes the device’s clock date and time.

ip: Define all the ip commands

ip address: Assigns an IP address and netmask to the selected interface.

ip route: Define the static ip routes

ip ospf: Assigns OSPF-specific parameters to an interface, such as the OSPF cost or authentication settings.

ip access-list: Used to create or modify an IP access list, specifying the version (standard or extended) and the list of rules to filter IP packets based on source/destination addresses, protocols, and ports.

ip domain-name: Sets the domain name for the device, which is used in various operations such as DNS resolution.

shutdown: Disable a router’s interface

no: Execute the opposite of the commands

no shutdown: Enable a router’s interface

no ntp server: Disable NTP

vlan: Define vlans. This will enter the Vlan Mode

name: Define the name of the vlan

`state: Define the state of the valn

switchport: Defines the switchports

router ospf: Configures and enables an OSPF routing process on the router. Specify the process ID to distinguish between multiple OSPF instances. This will enter the RouterConfig Mode

network: Associates a network or subnet with a specific OSPF area.

neighbor: Manually specifies a neighboring router for OSPF adjacency, usually in cases of non-broadcast networks.

area: Defines OSPF area-specific configurations, such as authentication, stub area settings, or default-cost for stub areas.

passive-interface: Prevents OSPF from sending hello packets on the specified interface while still advertising the interface’s network in OSPF.

distance: Configures the administrative distance for OSPF routes, which influences route preference when multiple protocols advertise the same destination.

default-information: Configures OSPF to advertise a default route (0.0.0.0/0) to other routers in the network.

router-id: Manually sets a unique identifier for the OSPF process, typically an IPv4 address, to distinguish the router in the OSPF domain.

clear ip ospf process: Restarts the OSPF process, clearing the OSPF routing table and adjacencies.

access-list: Defines an ACL by creating or modifying an access control list with a specified number or name. This command is used to specify a set of rules for filtering network traffic.

permit: An ACL action that allows network traffic that matches the rule’s conditions (e.g., specific IP address or protocol) to pass through.

deny: An ACL action that blocks network traffic matching the rule’s conditions, preventing it from passing through the network.

crypto: Defined all thr crypto commands

crypto ipsec profile: Configures and manages IPSec VPN profiles, including settings for security associations and tunnel configurations.

crypto key: Generates or manages cryptographic keys used in various security protocols, including VPNs and encryption.

set tranform-set: Specifies the transform set used in an IPSec VPN to define the cryptographic algorithms for encryption and integrity.

tunnel: Defines and manages the settings for an IPsec tunnel, including the associated transport and security protocols.

virtual-template: Creates a virtual template interface that can be used as a blueprint for creating virtual access interfaces, often used in VPN configurations.

ntp: Defines all the ntp commands

ntp server: Configures the NTP server for synchronizing time on the device, ensuring that the device’s clock is accurate.

ntp master: Configures the device as an NTP master, meaning it will serve time to other devices in the network.

ntp authenticate: Enables NTP authentication, which allows the NTP client to authenticate time synchronization requests from servers.

ntp authentication-key: Defines the key used for authenticating NTP messages, providing security to NTP transactions.

ntp trusted-key: Specifies which authentication key(s) are trusted to authenticate NTP messages

service password-encryption: Enables password encryption for storing sensitive passwords in the device’s configuration, ensuring they are not stored in plain text.

ping: Confirms the connection between ip addresses

Returns

A HashMap where the keys are command names (as &'static str) and the values are the corresponding Command structs. Each Command struct contains the name, description, suggestions, and an execute function.