pnf_cli::cliconfig
Struct CliContext

pub struct CliContext {

    pub current_mode: Mode,
    pub prompt: String,
    pub config: CliConfig,
    pub selected_interface: Option<String>,
    pub selected_vlan: Option<String>,
    pub vlan_names: Option<HashMap<String, String>>,
    pub vlan_states: Option<HashMap<u16, String>>,
    pub switchport_mode: Option<String>,
    pub trunk_encapsulation: Option<String>,
    pub native_vlan: Option<u16>,
    pub allowed_vlans: HashSet<u16>,
    pub ntp_servers: HashSet<String>,
    pub ntp_associations: Vec<NtpAssociation>,
    pub ntp_authentication_enabled: bool,
    pub ntp_authentication_keys: HashMap<u32, String>,
    pub ntp_trusted_keys: HashSet<u32>,
    pub ntp_master: bool,

}

Represents the current context of the CLI application.

The CliContext maintains the state of the CLI, including the current operational mode, the system prompt, the configuration, and the currently selected interface (if any).

Examples

let context = CliContext::default();
assert_eq!(context.prompt, "Router>");

Fields
current_mode: Modeprompt: Stringconfig: CliConfigselected_interface: Option<String>selected_vlan: Option<String>vlan_names: Option<HashMap<String, String>>vlan_states: Option<HashMap<u16, String>>switchport_mode: Option<String>trunk_encapsulation: Option<String>native_vlan: Option<u16>allowed_vlans: HashSet<u16>ntp_servers: HashSet<String>ntp_associations: Vec<NtpAssociation>ntp_authentication_enabled: boolntp_authentication_keys: HashMap<u32, String>ntp_trusted_keys: HashSet<u32>ntp_master: bool