pnf_cli::clock_settings
Function handle_show_clock

pub fn handle_show_clock(clock: &Clock)

Handles the show clock command to display the current date and time stored in the CustomClock structure.

Arguments

clock: A reference to the CustomClock instance whose date and time are to be displayed.

Usage

let clock = CustomClock {
    date: "2024-06-01".to_string(),
    time: "12:00".to_string(),
};
handle_show_clock(&clock);
// Output: Current clock: 2024-06-01 12:00