pnf_cli::commandcompleter
Struct CommandCompleter

pub struct CommandCompleter {
    pub commands: HashMap<String, Vec<String>>,
    pub current_mode: Mode,
}

A custom completer for the CLI application.

The CommandCompleter provides suggestions for commands based on user input. It integrates with the rustyline crate to offer real-time command-line assistance.

Fields

commands: A vector of strings containing the list of available commands.

current_mode: Gets the current mode of the cli

Fields
commands: HashMap<String, Vec<String>>current_mode: Mode
Implementations

impl CommandCompleter

Implementation of the CommandCompleter struct.

Provides a constructor for creating a new CommandCompleter instance, which is responsible for handling command completion based on the available commands and the current CLI mode.

pub fn new(commands: HashMap<String, Vec<String>>, current_mode: Mode) -> Self

Creates a new CommandCompleter instance.

Arguments

commands: A HashMap where the keys are command names (as String) and the values are vectors of possible completions or subcommands associated with each command.

current_mode: The current CLI mode (Mode), which influences the available commands and their completion behavior.

Returns

A new instance of CommandCompleter initialized with the provided commands and current mode.

Example

use std::collections::HashMap;
 
let mut commands = HashMap::new();
commands.insert("show".to_string(), vec!["ip".to_string(), "version".to_string()]);
commands.insert("configure".to_string(), vec!["terminal".to_string()]);
 
let completer = CommandCompleter::new(commands, Mode::UserMode);