pnf_cli::run_config
Function load_config

pub fn load_config() -> CliConfig

Loads the configuration from the startup-config.json file.

This function attempts to read the startup-config.json file and deserialize its contents into a CliConfig object. If the file cannot be opened, read, or parsed, a default configuration will be returned.

Returns

The function returns a CliConfig object. If loading the configuration fails, it will return the default configuration as defined by CliConfig::default().

Example

let config = load_config();
println!("Loaded config: {:?}", config);